adds a functional test for centos7 shaman repos
This test will create 3 vagrant vms. A shaman repo for ceph-medic will
be installed on the client0 node. The rest of the nodes will be used
by ceph-ansible to create a ceph cluster. That cluster is then tested by
ceph-medic from the client0 node.

The goal is to run this test nightly with jenkins.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>