debian: Build-Depends: python-tambo
Add python-tambo to the debian package's build deps, because the test
suite requires it now.

This also fixes a problem with the automatic dependencies from dh_python2:

dh_python2 automatically adds dependencies to the Debian package's
${python:Depends}. It searches the dependencies in .egg-info's
requires.txt and searches the installed Debian packages for a match.

Prior to this commit, dh_python2 could not find the Debian package name
for "tambo". From pbuilder's log:

  ... snip ...
  D: dh_python2 depends:107: generating dependencies for package ceph-medic
  D: dh_python2 pydist:121: trying to find dependency for execnet (python=2.7)
  D: dh_python2 pydist:121: trying to find dependency for tambo (python=2.7)
  D: dh_python2 pydist:161: invoking dpkg -S */[Tt][Aa][Mm][Bb][Oo]-?*\.egg-info | grep '/python2\../\|/pyshared/'
  D: dh_python2 pydist:177: dpkg -S did not find package for tambo: b'dpkg-query: no path found matching pattern */[Tt][Aa][Mm][Bb][Oo]-?*.egg-info\n'
  I: dh_python2 pydist:184: Cannot find package that provides tambo. Please add package that provides it to Build-Depends or add "tambo python-tambo" line to debian/pydist-overrides or add proper  dependency to Depends by hand and ignore this info.

When we add python-tambo to Build-Depends, pbuilder installs it in the
chroot, and dh_python2 auto-discovers it and adds it into
${python:Depends}.